Allow hue fixups in Bezier and expose interpolation as a plugin
